addplayer

+++ backend/baseball_pj/src/main/java/com/baseball/baseball_pj/Player/DTO/PlayerAdminRequestDTO.java
@@ -12,9 +12,9 @@ public class PlayerAdminRequestDTO {
     private Long playerId;
     private String playerName;
     private String playerPosition;
-    private Long playerBackNumber;
+    private Integer playerBackNumber;
     private LocalDate playerBirthDate;
     private String playerHeightWeight;
     private String playerEducationPath;
-    private Long teamId;
+    private Integer teamId;
 }
